refactor: Refactored BlogList component and Blog page to remove unused imports and functions, and updated code to use new BlogCard component.
